% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cluster_ari}
\alias{cluster_ari}
\title{Compute clustering ARI}
\usage{
cluster_ari(C_true, C_post)
}
\arguments{
\item{C_true}{True cluster assignemnts.}

\item{C_post}{Posterior responsibilities of predicted cluster assignemnts.}
}
\value{
The clustering ARI.
}
\description{
\code{cluster_ari} computes the clustering performance in terms
  of the Adjusted Rand Index (ARI) metric.
}
