% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cluster_error}
\alias{cluster_error}
\title{Compute clustering assignment error

\code{cluster_error} computes the clustering assignment error, i.e. the
average number of incorrect cluster assignments: \deqn{OE =
\sum_{n=1}^{N}(I(LT_{n} \neq LP_{n})) / N}}
\usage{
cluster_error(C_true, C_post)
}
\arguments{
\item{C_true}{True cluster assignemnts.}

\item{C_post}{Posterior mean of predicted cluster assignemnts.}
}
\value{
The clustering assignment error
}
\description{
Compute clustering assignment error

\code{cluster_error} computes the clustering assignment error, i.e. the
average number of incorrect cluster assignments: \deqn{OE =
\sum_{n=1}^{N}(I(LT_{n} \neq LP_{n})) / N}
}
