% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{eval_imputation_performance}
\alias{eval_imputation_performance}
\title{Evaluate imputation performance}
\usage{
eval_imputation_performance(obj, imputation_obj)
}
\arguments{
\item{obj}{Output of Melissa inference object.}

\item{imputation_obj}{List containing two vectors of equal length,
corresponding to true methylation states and predicted/imputed methylation
states.}
}
\value{
The `melissa` object, with an additional slot named `imputation`,
  containing the AUC, F-measure, True Positive Rate (TPR) and False Positive
  Rate (FPR), and Precision Recall (PR) curves.
}
\description{
\code{eval_imputation_performance} is a wrapper function for
  computing imputation/clustering performance in terms of different metrics,
  such as AUC and precision recall curves.
}
