% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_y}
\alias{extract_y}
\title{Extract responses y}
\usage{
extract_y(X, coverage_ind)
}
\arguments{
\item{X}{Observations}

\item{coverage_ind}{Which observations have coverage}
}
\value{
The design matrix H
}
\description{
Given a list of observations, extract responses y
}
