% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filter_by_coverage_across_cells}
\alias{filter_by_coverage_across_cells}
\title{Filter regions by coverage across cells}
\usage{
filter_by_coverage_across_cells(obj, min_cell_cov_prcg = 0.5)
}
\arguments{
\item{obj}{Melissa data object.}

\item{min_cell_cov_prcg}{Threshold on the proportion of cells that have
coverage for each region.}
}
\value{
Melissa object without the filtered genomic regions.
}
\description{
Filter genomic regions that have no coverage across many cells.
  This way we keep regions from which we can share information across cells.
}
