% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filter_by_cpg_coverage}
\alias{filter_by_cpg_coverage}
\title{Filter regions by coverage}
\usage{
filter_by_cpg_coverage(obj, min_cpgcov = 10)
}
\arguments{
\item{obj}{Melissa data object.}

\item{min_cpgcov}{Minimum CpG coverage for each genomic region.}
}
\value{
The same object with the regions that do not pass threshold set to
  NA.
}
\description{
Filter genomic regions that have low coverage within that
  region. This step does not remove the region, it only sets NA to those
  regions.
}
