% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filter_by_variability}
\alias{filter_by_variability}
\title{Filter regions by variability}
\usage{
filter_by_variability(obj, min_cov = 10)
}
\arguments{
\item{obj}{Melissa data object.}

\item{min_cov}{Minimum variability of mean methylation across cells, measured
in terms of standard deviation.}
}
\value{
Melissa object without the filtered genomic regions.
}
\description{
Filter genomic regions that have low mean methylation variability
  across cells; since they are not informative for cell subtype
  identification.
}
