% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{impute_met_state}
\alias{impute_met_state}
\title{Impute/predict methylation states}
\usage{
impute_met_state(obj, test, basis = NULL, is_predictive = TRUE,
  return_test = FALSE)
}
\arguments{
\item{obj}{Output of Melissa inference object.}

\item{test}{Test data to evaluate performance.}

\item{basis}{Basis object, if NULL we perform imputation using Melissa,
otherwise using BPRMeth.}

\item{is_predictive}{Logical, use predictive distribution for imputation, or
choose the cluster label with the highest responsibility.}

\item{return_test}{Whether or not to return a list with the predictions.}
}
\value{
A list containing two vectors, the true methylation state and the
  predicted/imputed methylation states.
}
\description{
Make predictions of missing methylation states, i.e. perfrom
  imputation using Melissa. This requires keepin a subset of data as a held
  out test set during Melissa inference.
}
