% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{init_design_matrix}
\alias{init_design_matrix}
\title{Initialise design matrices}
\usage{
init_design_matrix(basis, X, coverage_ind)
}
\arguments{
\item{basis}{Basis object.}

\item{X}{Observations}

\item{coverage_ind}{Which observations have coverage}
}
\value{
The design matrix H
}
\description{
Given a list of observations, initialise design matrices H for
  computational efficiency.
}
