% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{log_sum_exp}
\alias{log_sum_exp}
\title{Compute stable log-sum-exp}
\usage{
log_sum_exp(x)
}
\arguments{
\item{x}{A vector of observations}
}
\value{
The logs-sum-exp value
}
\description{
\code{log_sum_exp} computes the log sum exp trick for avoiding
  numeric underflow and have numeric stability in computations of small
  numbers.
}
\references{
\url{https://hips.seas.harvard.edu/blog/2013/01/09/computing-log-sum-exp/}
}
