% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Melissa_package.R
\docType{package}
\name{Melissa}
\alias{Melissa}
\alias{.datatable.aware}
\alias{Melissa-package}
\title{\code{Melissa}: Bayesian clustering and imputation of single cell
  methylomes}
\format{An object of class \code{logical} of length 1.}
\usage{
.datatable.aware
}
\value{
Melissa main package documentation.
}
\description{
Bayesian clustering and imputation of single cell methylomes
}
\seealso{
\code{\link{melissa}}, \code{\link{create_melissa_data_obj}},
  \code{\link{partition_dataset}}, \code{\link{plot_melissa_profiles}},
  \code{\link{filter_regions}}
}
\author{
C.A.Kapourani \email{kapouranis.andreas@gmail.com}
}
\keyword{datasets}
