% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Melissa_package.R
\docType{data}
\name{melissa_synth_dt}
\alias{melissa_synth_dt}
\title{Synthetic single cell methylation data}
\format{
A list object containing methylation regions, annotation data and the
  options used for creating the data. This in general would be the output of
  the \code{\link{create_melissa_data_obj}} function. It has the following
  three objects: \itemize{ \item{ \code{met}: A list containing the
  methylation data, each element of the list is a different cell.} \item{
  \code{anno_region}: Corresponding annotation data for each genomic region.}
  \item{ \code{opts}: Parameters/options used to generate the data.} }
}
\usage{
melissa_synth_dt
}
\value{
Synthetic methylation data
}
\description{
Small synthetic data for quick analysis. It consists of N = 50
  cells and M = 50 genomic regions.
}
\seealso{
\code{\link{create_melissa_data_obj}}
}
\keyword{datasets}
