% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plot.R
\name{plot_melissa_profiles}
\alias{plot_melissa_profiles}
\title{Plot predictive methylaation profiles}
\usage{
plot_melissa_profiles(melissa_obj, region = 1,
  title = "Melissa profiles", x_axis = "genomic region",
  y_axis = "met level", x_labels = c("Upstream", "", "Centre", "",
  "Downstream"), ...)
}
\arguments{
\item{melissa_obj}{Clustered cell subtypes using Melissa inference functions.}

\item{region}{Genomic region number.}

\item{title}{Plot title}

\item{x_axis}{x axis label}

\item{y_axis}{x axis label}

\item{x_labels}{x axis ticks labels}

\item{...}{Additional parameters}
}
\value{
A ggplot2 object.
}
\description{
This function plots the predictive distribution of the
  methylation profiles inferred using the Melissa model. Each colour
  corresponds to a different cluster.
}
\author{
C.A.Kapourani \email{C.A.Kapourani@ed.ac.uk}
}
