% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{eval_cluster_performance}
\alias{eval_cluster_performance}
\title{Evaluate clustering performance}
\usage{
eval_cluster_performance(obj, C_true)
}
\arguments{
\item{obj}{Output of Melissa inference object.}

\item{C_true}{True cluster assignemnts.}
}
\value{
The `melissa` object, with an additional slot named `clustering`,
  containing the ARI and clustering assignment error performance.
}
\description{
\code{eval_cluster_performance} is a wrapper function for
  computing clustering performance in terms of ARI and clustering assignment
  error.
}
